third_party/blink/web_tests/webaudio/OfflineAudioContext/offlineaudiocontext-constructor.html

<!DOCTYPE html>
<html>
  <head>
    <title>offlineaudiocontext-constructor.html</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(t => {
        let context;
        // Make sure we don't crash when giving 0 as number of frames.
        assert_throws_dom(
            'NotSupportedError',
            () => new OfflineAudioContext(1, 0, 44100),
            'new OfflineAudioContext(1, 0, 44100)');
        // Make sure we don't throw exceptions for supported ranges of
        // sample rates for an OfflineAudioContext.
        assert_throws_dom(
            'NotSupportedError',
            () => context = new OfflineAudioContext(2, 512, 2999),
            'context0 = new OfflineAudioContext(2, 512, 2999)');

        context = new OfflineAudioContext(2, 512, 3000);
        assert_equals(context.length, 512, 'context1.length');

        context = new OfflineAudioContext(2, 1024, 768000);
        assert_equals(context.length, 1024, 'context2.length');

        assert_throws_dom(
            'NotSupportedError',
            () => new OfflineAudioContext(2, 1024, 768001),
            'context3 = new OfflineAudioContext(2, 1024, 768001)');

        context = new OfflineAudioContext(2, 2048, 8000);
        assert_equals(context.length, 2048, 'context4.length');

        context = new OfflineAudioContext(2, 4096, 11025);
        assert_equals(context.length, 4096, 'context5.length');

        context = new OfflineAudioContext(2, 512, 22050);
        assert_equals(context.length, 512, 'context6.length');

        context = new OfflineAudioContext(2, 512, 44100);
        assert_equals(context.length, 512, 'context7.length');

        context = new OfflineAudioContext(2, 512, 48000);
        assert_equals(context.length, 512, 'context8.length');

        context = new OfflineAudioContext(2, 512, 88200);
        assert_equals(context.length, 512, 'context9.length');
        // Make sure length is read-only.
        context.length = 99;
        assert_equals(
            context.length, 512, 'context9.length after illegal setter value');

        context = new OfflineAudioContext(2, 512, 96000);

        // Make sure we throw exceptions for non-finite sample rates.
        assert_throws_js(
            TypeError,
            () => new OfflineAudioContext(1, 0, NaN),
            'context11 = new OfflineAudioContext(1, 0, NaN)');
        assert_throws_js(
            TypeError,
            () => new OfflineAudioContext(1, 0, Infinity),
            'context12 = new OfflineAudioContext(1, 0, Infinity)');

        // Verify channel counts and other destination attributes
        // are set correctly.
        context = new OfflineAudioContext(7, 512, 48000);

        assert_equals(
            context.destination.channelCount,
            7,
            'context13.destination.channelCount');
        assert_equals(
            context.destination.maxChannelCount,
            7,
            'context13.destination.maxChannelCount');
        assert_equals(
            context.destination.channelCountMode,
            'explicit',
            'context13.destination.channelCountMode');
        assert_equals(
            context.destination.channelInterpretation,
            'speakers',
            'context13.destination.channelInterpretation');
      }, 'OfflineAudioContext constructor behavior validation');
    </script>
  </body>
</html>
